refactor: :truck: update name && import Footer component from mantine
